fix: documentation build
