+++ packages/documentation/next.config.mjs
@@ -11,6 +11,8 @@ const nextConfig = {
   turbopack: {
     resolveAlias: {
       "next-mdx-import-source-file": "./src/mdx-components.tsx",
+      "@nahkies/openapi-code-generator/web":
+        "./node_modules/@nahkies/openapi-code-generator/dist/cjs/web.cjs",
     },
   },
 }
